Standalone Updates
Updates to asei.js were made to allow a more open environment to use the
script at different levels of integration. The standalone changes allow
the script to be used by itself on any web-server and with any
server-side language but will require more work for implementation.

The documentation of the code was updated to reflect these changes.

The Readme was also updated with possible configuration options and some
general informational updates.